// openapi/collector/model/ctListSuperQueryInfo.ts
/**
 * 
 * MPDC Collector API
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { QueryData } from './queryData';
import { CtListQueryOptions } from './ctListQueryOptions';
import { CtListSubQueryInfo } from './ctListSubQueryInfo';


export interface CtListSuperQueryInfo { 
    queryData: QueryData;
    listData: CtListQueryOptions;
    fetchMultimedia: Array<CtListSuperQueryInfo.FetchMultimediaEnum>;
    indexInCineast: boolean;
    subQueries: Array<CtListSubQueryInfo>;
    id: number;
}
export namespace CtListSuperQueryInfo {
    export type FetchMultimediaEnum = 'UNKNOWN' | 'ANIMATED_IMAGE' | 'IMAGE' | 'VIDEO';
    export const FetchMultimediaEnum = {
        Unknown: 'UNKNOWN' as FetchMultimediaEnum,
        AnimatedImage: 'ANIMATED_IMAGE' as FetchMultimediaEnum,
        Image: 'IMAGE' as FetchMultimediaEnum,
        Video: 'VIDEO' as FetchMultimediaEnum
    };
}